Rename init1 to base_mesh and reorganize
A new config file has been added and a python script now
performs all the steps of setting up and culling the base mesh.
Some namelist options can be provided in the config file and will
be substituted at runtime before running the model.

Geometry processing that been reorganized to be in a function
call, allowing it to be called from the base_mesh.py script
(not as a subprocess).